refactor: move photo to provider dir
